Added extra field to interface to add extra variables. Added functionality to MACD, and also a series to count number of ticks

+++ Simple Sim/UtilityFunctions.h
@@ -17,8 +17,8 @@
 +(int) signOfInt: (int) n;
 
 +(void) calcSortIndexForDoubleArray: (double *) arrayToSort
-                     WithStartIndex: (NSUInteger) startIndex
-                        AndEndIndex: (NSUInteger) endIndex
+                     WithStartIndex: (long) startIndex
+                        AndEndIndex: (long) endIndex
               AndReturningSortIndex: (int *) sortIndexArray;
 
 @end
